Fix #132700: Grease Pencil: Strokes with locked materials receive weight value
With this fix strokes with locked materials are not affected any more
by the Grease Pencil weight paint tools. The vertex weights in
material-locked strokes are read-only: they are not changed, but the
weights do count for Average, Blur and Smear.

Pull Request: https://projects.blender.org/blender/blender/pulls/138900