Mesh: Separate GPU subdivision normals and positions
Currently for the result of GPU subdivision, normals are stored in two
places. First, interleaved with the positions, and second, in a separate
"lnor" vertex buffer. This extra duplication costs 12 bytes of GPU
memory usage per subdivided face corner, which is quite a lot. It also
causes problems because it's inconsistent with the format used for
regular non-subdivided meshes, causing various hacks and issues
throughout the code.

This PR moves the normals into just the "lnor" vertex buffer. This
starts some code simplifications, but ends up still more complicated
than I'd like because some edit selection flags are still stored in the
fourth component of the normal vector. Tacking that makes this PR too
big, but it would be a good next step.

Pull Request: https://projects.blender.org/blender/blender/pulls/137158