Assets: Remove asset handle RNA type
The `AssetHandle` RNA type was added to enable the asset view template,
which was removed in ae9ca35e3b. The type itself was hacky and always
considered a temporary design, as suggested in the type's RNA
description. It has by now been superseded by the `AssetRepresentation`
type.

This is another important step towards removing `AssetHandle`, see
blender/blender#108806.