UI: Properly disable the Remove Material Slot Button in edit mode
Currently the remove material slot button still appears to be
active while in edit mode. Attempting to use the button will
result in an error "Unable to remove material slot in edit mode."

Solution is to edit poll function to disable/gray out the remove
material slot button while in edit mode similar to other buttons

In the code a check was added for edit mode for the operator
execution as a fix for bug #21822. However this check was never added
to the polling function. This commit adds that check.

Pull Request: https://projects.blender.org/blender/blender/pulls/139660